requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pyyaml==6.0.1
apscheduler==3.10.4
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dateutil==2.9.0
pytz==2024.1
twilio==9.1.0          # only needed if using Twilio provider
rich==13.7.1           # pretty log viewer
pytest==8.2.0
pytest-asyncio==0.23.6
